Fixed LibraryEntriesRemoteMediator is created repeatedly causing it not refreshing
RemoteMediator was not initially refreshing if the start page is set to Library.

This was caused because the dataSource flow was emitting multiple times, even if the filter has not changed.
We are now explicitly calling distinctUntilChanged() on the flow and the LibraryEntryFilter data class is no longer mutated when pageSize() is set.