Fixed searching in library not working
Before the change, we added the search query filter to the LibraryEntryFilter object inside the flapMap of the data source flow. Since we added distinctUntilChanged() right before the flapMap, the flow is never updated.

To fix this, we directly add the search query filter to the LibraryEntryFilter object when it changes, so distinctUntilChanged() will pass and the flow gets updated.