Fixed crash in LibraryEntriesAdapter
The getItemViewType(pos) method returns the resource ID of the layout that should be displayed on the given position. To determine the item type on the position, we use peek(pos) and if the returned type is unknown we have thrown an exception.
However, the item returned by peek(pos) can be null, for example when canceling an ongoing library search.

To fix the crashes, we now return `0` for the view type (like the default implementation of getItemViewType does).